refactor(structure): rename components/cosmos to components/ui
Aligns the design system folder with the standard ui/ convention.
Updates all TypeScript imports and CSS composes references across the codebase.
