+++ src/components/ui/background/draw.ts
