+++ src/components/ui/background/simulation.ts
